+++ eds4jinja2/adapters/__init__.py
@@ -4,7 +4,7 @@
 # Date:  07/08/2020
 # Author: Eugeniu Costetchi
 # Email: costezki.eugen@gmail.com
-
+import collections
 from typing import List, Dict
 
 
@@ -62,3 +62,19 @@ def invert_dict(mapping_dict: Dict, reduce_values: bool = True):
     if reduce_values:
         return {k: sort_by_size_and_alphabet(v)[0] for k, v in inv_map.items()}
     return inv_map
+
+
+def deep_update(source, overrides):
+    """
+    Update a nested dictionary or similar mapping.
+    Modify ``source`` in place.
+
+    Used from https://stackoverflow.com/questions/3232943/update-value-of-a-nested-dictionary-of-varying-depth
+    """
+    for key, value in overrides.items():
+        if isinstance(value, collections.Mapping) and value:
+            returned = deep_update(source.get(key, {}), value)
+            source[key] = returned
+        else:
+            source[key] = overrides[key]
+    return source
