+++ tests/unit/test_namespace_handlers.py
@@ -105,4 +105,3 @@ def test_new_namespace_inventory(dummy_prefixes):
     assert nm.uri_to_qname("http://www.w3.org/2004/02/skos/core#Concept") == "skos:Concept"
 
 
-
